src/scheduler/preprocess/requirements.txt
scikit-learn==0.22.2.post1
pandas==0.25.1
numpy==1.19.4
click
PyYAML==6.0
marshmallow-dataclass==8.5.8
marshmallow==3.17.0
